move all view Options struct to Options.h
View now uses aliases to these top level structs
